Task:  Election Polls
Task:  Election Polls
In an elections, each voter writes the name of a candidate on a ballot and places it in the ballot box. The candidate with the highest number of votes wins the election. If two or more candidates have the same number of votes, then the tied candidates' names are ordered alphabetically and the last name in the alphabetical order wins. For example, votes are in the names ['Jessy', 'Philip', 'Jessy', 'Philip']. Each candidate received two votes, but Philip is alphabetically later than Jessy, so he wins.