Task:  Let's sort Question 2
Task:  Let's sort
You will be given an array of integers to sort. Sorting must first be by frequency of occurrence, then by value. For instance, given an array [4, 5, 7, 7, 6, 5, 4, 3], there is one each of 6's and 3's, and there are two 4's, two 7's and two 5's. The sorted list is [3, 6, 4, 4, 5, 5, 7, 7].