Task:  Get rid of duplicates:: Question 3
Task:  Get rid of duplicates
Given an array of integers, your task is to count the number of duplicate array elements and get rid of them to form a new array. Duplicate is defined as two or more identical elements. For example, in the array [4, 5, 5, 6, 6, 6], the two fives are one duplicate and so are the three sixes.