Task:  Sticks Cutting:: Question 5
Task:  Sticks Cutting
You will be given an array of the lengths of a number of sticks. When a turn begins, you must count the number of sticks you have. Determine the length of a stick with the shortest length and discard any sticks of that length. Remove that length from each of the longer sticks and discard the offcuts. Repeat until there are no sticks left. Return an array where the elements are the numbers of sticks you had at the beginning of each turn. For example, consider an array representing the lengths of four sticks: [1, 1, 2, 3]. The shortest sticks are 1 unit long. Discard them. Remove 1 unit from the other two sticks and discard the scrap. Now you have two sticks lengths [1, 2]. Do the same and you'll have one stick of length [1]. Discard it and return an array with the number of sticks you had at the start of each turn: [4, 2, 1].