Task:  Remove Redundancy:: Question 6
Task:  Remove Redundancy
A redundant node in a singly-linked list is defined as a node whose value matches the value of a previous node in the list. In other words, an earlier value recurs. Given a reference to the head of a singly-linked list that may contain redundant nodes, create a singly-linked list that contains only unique values as they first occur in the inputs.